Do not delete HTTP host header when making request to localhost
45c02705 was intended as a fix for the case where a request to

		local.hubspot.com

is proxied to a remote server like

		app.hubspot.com

In that case, not deleting the Host header causes the request to be sent
with "Host: local.hubspot.com", which doesn't make any sense.

However, this introduced an issue with requests to localhost: Some local
services behave differently depending on the host name, such as
switching between QA and prod mode based on whether the request is made
to local.hubspotqa.com or local.hubspot.com. Therefore, this commit
makes vee leave the host header as-is if the proxy target is localhost.